Remove modules kas-core::event::cx::{cx_pub,platform}
